PS-Caffiene
My take on keeping my systems awake. I provided my PS1 file and Compiled EXE (Sapien, which I have via work). The Sapien EXE runs in the taskbar and can be right-clicked and exited, closing the loop. You can also call up EXE and use -duration X and specify the number of hours to run it. The default is 0 and is run until you exit it.

Enjoy and if I need features or bug fixes, let me know. The EXE should be signed with a trusted cert and makes it easier to click and run.